Create read_cell_append_cell.py
This reads cells from an excel document creates a dictionary from this data and uses the keys from the dictionaries to go through another excel document and match the sheetname to the dictionary key.  After this matchup the program then takes all the values associated with that key that matched the sheetname and starts searching through each row from the bottom up and inserts this data below the latest data entered.  It still needs a function to where if the sheetname isn't matched then it creates a new sheet.